+++ hypha/server.py
@@ -16,7 +16,6 @@ from hypha.core.auth import create_login_service
 from hypha.core.store import RedisStore
 from hypha.core.queue import create_queue_service
 from hypha.http import HTTPProxy
-from hypha.log import EventLoggingService
 from hypha.triton import TritonProxy
 from hypha.utils import GZipMiddleware, GzipRoute, PatchedCORSMiddleware
 from hypha.websocket import WebsocketServer
@@ -84,9 +83,6 @@ def start_builtin_services(
     store.register_public_service(create_queue_service(store))
     store.register_public_service(create_login_service(store))
 
-    if args.database_uri:
-        EventLoggingService(store, database_uri=args.database_uri)
-
     if args.enable_s3:
         # pylint: disable=import-outside-toplevel
         from hypha.s3 import S3Controller
@@ -108,7 +104,6 @@ def start_builtin_services(
             store,
             s3_controller=s3_controller,
             workspace_bucket=args.workspace_bucket,
-            database_uri=args.database_uri,
         )
 
     if args.enable_server_apps:
@@ -225,6 +220,7 @@ def create_application(args):
         public_base_url=public_base_url,
         local_base_url=local_base_url,
         redis_uri=args.redis_uri,
+        database_uri=args.database_uri,
         reconnection_token_life_time=float(
             env.get("RECONNECTION_TOKEN_LIFE_TIME", str(2 * 24 * 60 * 60))
         ),